# Conflicts:
#	nts.uk/uk.com/com.ctx/bs/person/nts.uk.ctx.bs.person.infra/src/main/java/repository/person/info/setting/JpaPerInfoInitValSetItem.java

+++ nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.dom/src/main/java/nts/uk/ctx/bs/employee/dom/jobtitle/info/SequenceMasterRepository.java
@@ -56,6 +56,5 @@ public interface SequenceMasterRepository {
 	 *
 	 * @return the short
 	 */
-	short findMaxOrder();
-	
+	int findMaxOrder();
 }
