+++ nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.infra/src/main/java/nts/uk/ctx/bs/employee/infra/repository/jobtitle/JpaSequenceMasterSetMemento.java
@@ -36,7 +36,7 @@ public class JpaSequenceMasterSetMemento implements SequenceMasterSetMemento {
 	 * SequenceMasterSetMemento#setOrder(int)
 	 */
 	@Override
-	public void setOrder(short order) {
+	public void setOrder(int order) {
 		this.typeValue.setDisporder(order);
 	}
